Move the bank menu to its own class
